# pizza-rendeles-client/src/swagger/model/personalDetailsWrapper.ts
/**
 * REST API SPEC FOR MIX WEBAPP
 * REST API SPEC FOR MIX WEBAPP
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { PersonalDetails } from './personalDetails';


export interface PersonalDetailsWrapper { 
    personalDetails?: Array<PersonalDetails>;
}