# src/main/resources/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "REST API SPEC FOR MIX WEBAPP",
    "version" : "v1",
    "title" : "REST API SPEC FOR MIX WEBAPP"
  },
  "tags" : [ {
    "name" : "BoxedPizza"
  }, {
    "name" : "Order"
  }, {
    "name" : "PersonalDetails"
  }, {
    "name" : "Pizza"
  }, {
    "name" : "Sauce"
  }, {
    "name" : "Topping"
  } ],
  "paths" : {
    "/api/boxedPizza/pizzas/{pizzaID}/boxedpizzas" : {
      "get" : {
        "tags" : [ "BoxedPizza" ],
        "summary" : "Get boxed pizzas",
        "description" : "",
        "operationId" : "getBoxedPizzas",
        "parameters" : [ {
          "name" : "pizzaID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/BoxedPizzaWrapper"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "BoxedPizza" ],
        "summary" : "Add boxed pizza",
        "description" : "",
        "operationId" : "addBoxedPizza",
        "parameters" : [ {
          "name" : "pizzaID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/BoxedPizza"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/BoxedPizza"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "BoxedPizza" ],
        "summary" : "Delete boxed pizza",
        "description" : "",
        "operationId" : "deleteBoxedPizza",
        "parameters" : [ {
          "name" : "pizzaID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "boxID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object"
            }
          }
        }
      }
    },
    "/api/boxedPizza/pizzas/{pizzaID}/boxedpizzas/{boxID}" : {
      "put" : {
        "tags" : [ "BoxedPizza" ],
        "summary" : "Update boxed pizza",
        "description" : "",
        "operationId" : "updateBoxedPizza",
        "parameters" : [ {
          "name" : "pizzaID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "boxID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/BoxedPizza"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/BoxedPizza"
            }
          }
        }
      }
    },
    "/api/order/people/{personID}/orders" : {
      "get" : {
        "tags" : [ "Order" ],
        "summary" : "Get orders",
        "description" : "",
        "operationId" : "getOrders",
        "parameters" : [ {
          "name" : "personID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/OrderWrapper"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Order" ],
        "summary" : "Add new order",
        "description" : "",
        "operationId" : "addOrder",
        "parameters" : [ {
          "name" : "personID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Order"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Order"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Order" ],
        "summary" : "Delete order",
        "description" : "",
        "operationId" : "deleteOrders",
        "parameters" : [ {
          "name" : "personID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "orderID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object"
            }
          }
        }
      }
    },
    "/api/order/people/{personID}/orders/{orderID}" : {
      "put" : {
        "tags" : [ "Order" ],
        "summary" : "Update order",
        "description" : "",
        "operationId" : "updateOrder",
        "parameters" : [ {
          "name" : "personID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "orderID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Order"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Order"
            }
          }
        }
      }
    },
    "/api/personalDetails/people" : {
      "get" : {
        "tags" : [ "PersonalDetails" ],
        "summary" : "Get all personal details",
        "description" : "",
        "operationId" : "getPersonalDetails",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PersonalDetailsWrapper"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "PersonalDetails" ],
        "summary" : "Add new personal details",
        "description" : "",
        "operationId" : "addPersonalDetails",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/PersonalDetails"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PersonalDetails"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "PersonalDetails" ],
        "summary" : "Delete personal details",
        "description" : "",
        "operationId" : "deletePersonalDetails",
        "parameters" : [ {
          "name" : "personID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object"
            }
          }
        }
      }
    },
    "/api/personalDetails/people/{personID}" : {
      "put" : {
        "tags" : [ "PersonalDetails" ],
        "summary" : "Update personal details",
        "description" : "",
        "operationId" : "updatePersonalDetails",
        "parameters" : [ {
          "name" : "personID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/PersonalDetails"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PersonalDetails"
            }
          }
        }
      }
    },
    "/api/pizza/pizzas" : {
      "get" : {
        "tags" : [ "Pizza" ],
        "summary" : "Get Pizza types",
        "description" : "",
        "operationId" : "getPizzaTypes",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PizzaWrapper"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Pizza" ],
        "summary" : "Add Pizza Type",
        "description" : "",
        "operationId" : "addPizza",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Pizza"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Pizza"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Pizza" ],
        "summary" : "Delete pizza",
        "description" : "",
        "operationId" : "deletePizza",
        "parameters" : [ {
          "name" : "pizzaID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object"
            }
          }
        }
      }
    },
    "/api/pizza/pizzas/{pizzaID}" : {
      "put" : {
        "tags" : [ "Pizza" ],
        "summary" : "Update Type",
        "description" : "",
        "operationId" : "updatePizza",
        "parameters" : [ {
          "name" : "pizzaID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Pizza"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Pizza"
            }
          }
        }
      }
    },
    "/api/sauce/sauces" : {
      "get" : {
        "tags" : [ "Sauce" ],
        "summary" : "Get sauces",
        "description" : "",
        "operationId" : "getSauces",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SauceWrapper"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Sauce" ],
        "summary" : "Add new sauce",
        "description" : "",
        "operationId" : "addSauces",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Sauce"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Sauce"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Sauce" ],
        "summary" : "Get sauces",
        "description" : "",
        "operationId" : "deleteSauces",
        "parameters" : [ {
          "name" : "sauceID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object"
            }
          }
        }
      }
    },
    "/api/sauce/sauces/{sauceID}" : {
      "put" : {
        "tags" : [ "Sauce" ],
        "summary" : "Update sauce",
        "description" : "",
        "operationId" : "updateSauces",
        "parameters" : [ {
          "name" : "sauceID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Sauce"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Sauce"
            }
          }
        }
      }
    },
    "/api/topping/pizzas/{pizzaID}/toppings" : {
      "get" : {
        "tags" : [ "Topping" ],
        "summary" : "Get all toppings",
        "description" : "",
        "operationId" : "getToppings",
        "parameters" : [ {
          "name" : "pizzaID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ToppingWrapper"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Topping" ],
        "summary" : "Add new topping",
        "description" : "",
        "operationId" : "addTopping",
        "parameters" : [ {
          "name" : "pizzaID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Topping"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Topping"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Topping" ],
        "summary" : "Delete topping",
        "description" : "",
        "operationId" : "deleteTopping",
        "parameters" : [ {
          "name" : "pizzaID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "toppingID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object"
            }
          }
        }
      }
    },
    "/api/topping/pizzas/{pizzaID}/toppings/{toppingID}" : {
      "put" : {
        "tags" : [ "Topping" ],
        "summary" : "Update topping",
        "description" : "",
        "operationId" : "updateTopping",
        "parameters" : [ {
          "name" : "pizzaID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "toppingID",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Topping"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Topping"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "BoxedPizza" : {
      "type" : "object",
      "properties" : {
        "pizzaID" : {
          "type" : "integer",
          "format" : "int32"
        },
        "size" : {
          "type" : "integer",
          "format" : "int32"
        },
        "price" : {
          "type" : "integer",
          "format" : "int32"
        },
        "pizza" : {
          "$ref" : "#/definitions/Pizza"
        },
        "thick" : {
          "type" : "boolean"
        }
      }
    },
    "BoxedPizzaWrapper" : {
      "type" : "object",
      "properties" : {
        "boxedPizzas" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/BoxedPizza"
          }
        }
      }
    },
    "Order" : {
      "type" : "object",
      "properties" : {
        "orderID" : {
          "type" : "integer",
          "format" : "int32"
        },
        "orderPrice" : {
          "type" : "integer",
          "format" : "int32"
        },
        "orderDate" : {
          "type" : "string"
        },
        "personalDetails" : {
          "$ref" : "#/definitions/PersonalDetails"
        },
        "boxedPizzas" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "$ref" : "#/definitions/BoxedPizza"
          }
        }
      }
    },
    "OrderWrapper" : {
      "type" : "object",
      "properties" : {
        "orders" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Order"
          }
        }
      }
    },
    "PersonalDetails" : {
      "type" : "object",
      "properties" : {
        "userID" : {
          "type" : "integer",
          "format" : "int32"
        },
        "personalName" : {
          "type" : "string"
        },
        "address" : {
          "type" : "string"
        },
        "phoneNumber" : {
          "type" : "string"
        },
        "emailAddress" : {
          "type" : "string"
        }
      }
    },
    "PersonalDetailsWrapper" : {
      "type" : "object",
      "properties" : {
        "personalDetails" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/PersonalDetails"
          }
        }
      }
    },
    "Pizza" : {
      "type" : "object",
      "properties" : {
        "pizzaName" : {
          "type" : "string"
        },
        "sauce" : {
          "$ref" : "#/definitions/Sauce"
        },
        "toppings" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Topping"
          }
        },
        "spicy" : {
          "type" : "boolean"
        }
      }
    },
    "PizzaWrapper" : {
      "type" : "object",
      "properties" : {
        "pizzas" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Pizza"
          }
        }
      }
    },
    "Sauce" : {
      "type" : "object",
      "properties" : {
        "sauceName" : {
          "type" : "string"
        },
        "color" : {
          "type" : "string"
        },
        "spicy" : {
          "type" : "boolean"
        }
      }
    },
    "SauceWrapper" : {
      "type" : "object",
      "properties" : {
        "sauces" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Sauce"
          }
        }
      }
    },
    "Topping" : {
      "type" : "object",
      "properties" : {
        "toppingName" : {
          "type" : "string"
        },
        "category" : {
          "type" : "string"
        },
        "isSpicy" : {
          "type" : "boolean"
        }
      }
    },
    "ToppingWrapper" : {
      "type" : "object",
      "properties" : {
        "toppings" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Topping"
          }
        }
      }
    }
  }
}